+++ modules/m_svsjoin.c
@@ -66,7 +66,7 @@ ms_svsjoin(struct Client *source_p, int parc, char *parv[])
 
   if (target_p->from == source_p->from)
   {
-    sendto_realops_flags(UMODE_DEBUG, L_ALL, SEND_NOTICE,
+    sendto_realops_flags(UMODE_SERVNOTICE, L_ALL, SEND_NOTICE,
                          "Received wrong-direction SVSJOIN "
                          "for %s (behind %s) from %s",
                          target_p->name, source_p->from->name,