+++ modules/m_svskill.c
@@ -77,7 +77,7 @@ ms_svskill(struct Client *source_p, int parc, char *parv[])
 
   if (target_p->from == source_p->from)
   {
-    sendto_realops_flags(UMODE_DEBUG, L_ALL, SEND_NOTICE,
+    sendto_realops_flags(UMODE_SERVNOTICE, L_ALL, SEND_NOTICE,
                          "Received wrong-direction SVSKILL "
                          "for %s (behind %s) from %s",
                          target_p->name, source_p->from->name,