+++ src/conf.c
@@ -627,7 +627,7 @@ conf_set_defaults(void)
   ConfigGeneral.ping_cookie = 0;
   ConfigGeneral.no_oper_flood = 0;
   ConfigGeneral.max_targets = MAX_TARGETS_DEFAULT;
-  ConfigGeneral.oper_only_umodes = UMODE_DEBUG | UMODE_LOCOPS | UMODE_HIDDEN | UMODE_FARCONNECT |
+  ConfigGeneral.oper_only_umodes = UMODE_LOCOPS | UMODE_HIDDEN | UMODE_FARCONNECT |
                                    UMODE_EXTERNAL | UMODE_FLOOD | UMODE_NCHANGE |
                                    UMODE_SPY | UMODE_SKILL | UMODE_REJ | UMODE_CCONN;
   ConfigGeneral.oper_umodes = UMODE_FLOOD | UMODE_LOCOPS | UMODE_SERVNOTICE | UMODE_WALLOP;