+++ src/s_bsd.c
@@ -112,9 +112,6 @@ void
 report_error(int level, const char *text, const char *who, int error)
 {
   who = (who) ? who : "";
-
-  sendto_realops_flags(UMODE_DEBUG, level, SEND_NOTICE,
-                       text, who, strerror(error));
   ilog(LOG_TYPE_IRCD, text, who, strerror(error));
 }